added security with Bearer Authentication (jwtoken) updated integration test with signup to get a jwtoken for authentication added integration test for TaskController.

+++ src/main/java/pb/se/TimeReportService/service/CustomerService.java
@@ -4,6 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pb.se.TimeReportService.domain.Customer;
 import pb.se.TimeReportService.domain.User;
+import pb.se.TimeReportService.exception.CustomerNotFoundException;
+import pb.se.TimeReportService.exception.ForbiddenException;
 import pb.se.TimeReportService.port.persistence.CustomerRepository;
 
 import java.util.List;
@@ -18,38 +20,38 @@ public class CustomerService {
     @Autowired
     private UserService userService;
 
-    public List<Customer> getAllCustomers() {
-        return customerRepository.findAll();
+    public Customer getCustomerById(User user, UUID id) {
+        validateUserIsTheOwner(user, id);
+        return customerRepository.findById(id).orElseThrow(CustomerNotFoundException::new);
     }
 
-    public Optional<Customer> getCustomerById(UUID id) {
-        return customerRepository.findById(id);
-    }
-
-    public Customer createCustomer(Customer customer) {
-        return customerRepository.save(customer);
-    }
-
-    public Optional<Customer> updateCustomer(UUID id, Customer customer) {
+    public Optional<Customer> updateCustomer(User user, UUID id, Customer customer) {
+        validateUserIsTheOwner(user, id);
         return customerRepository.findById(id)
                 .map(existingCustomer -> {
                     existingCustomer.setName(customer.getName());
+                    existingCustomer.setTasks(customer.getTasks());
                     return customerRepository.save(existingCustomer);
                 });
     }
 
-    public boolean deleteCustomer(UUID id) {
-        return customerRepository.findById(id)
-                .map(customer -> {
-                    customerRepository.delete(customer);
-                    return true;
-                })
-                .orElse(false);
+    public void deleteCustomer(UUID id, User user) {
+        validateUserIsTheOwner(user, id);
+        Customer customer = customerRepository.findById(id).orElseThrow(CustomerNotFoundException::new);
+
+        customerRepository.delete(customer);
     }
 
-    public void addCustomer(User user, Customer customer) {
-        customerRepository.save(customer);
+    public Customer addCustomer(User user, Customer customer) {
+        Customer saved = customerRepository.save(customer);
         user.addCustomer(customer);
         userService.update(user);
+        return saved;
     }
+    private static void validateUserIsTheOwner(User user, UUID id) {
+        if (user.getCustomers().stream().noneMatch(c-> c.getId().equals(id))) {
+            throw new ForbiddenException();
+        }
+    }
+
 }
